requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.2.2
playwright==1.44.0
python-dotenv==1.0.1
